Move the New Issue document projection into its own module
Summarising the form pushed issue_detail_content.rs past the 750-line
advisory ceiling the source-size gate enforces. The new-issue rows are a
self-contained projection, so give them a sibling module next to the
tests that already cover them and re-export the entry point.
